requirements.txt
streamlit>=1.30.0
playwright>=1.40.0
playwright-stealth>=1.0.6
openai>=1.10.0
anthropic>=0.18.0
openpyxl>=3.1.2
python-docx>=1.1.0
beautifulsoup4>=4.12.0
requests>=2.31.0
